Lots of tweaks

+++ Base/StringFmt.c
@@ -240,9 +240,9 @@ int vStringFmt(char *b, int size, const char *f, va_list args)
       if(print0x) {
 	field[0] = '0';
 	field[1] = 'x';
-	fieldLen = 2 + bufferPrintUL(&field[2], PRINT_FMT_FIELD-2, (unsigned long) va_arg(args, unsigned int), 16, p);
+	fieldLen = 2 + bufferPrintUL(&field[2], PRINT_FMT_FIELD-2, (unsigned long) va_arg(args, unsigned long), 16, p);
       } else
-	fieldLen = bufferPrintUL(field, PRINT_FMT_FIELD, (unsigned long) va_arg(args, unsigned int), 16, p);
+	fieldLen = bufferPrintUL(field, PRINT_FMT_FIELD, (unsigned long) va_arg(args, unsigned long), 16, p);
 	
       len += stringAppendPadded(&b[len], size - len, w, field, fieldLen, padChar, leftJust);
       break;
